requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
constantly==23.10.4
crontab==1.0.1
cryptography==43.0.3
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-rq==2.10.2
django-scheduler==0.10.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-hooks==0.1.3
drf-spectacular==0.27.2
environ==1.0
et_xmlfile==2.0.0
freezegun==1.5.1
hyperlink==21.0.0
icalendar==6.0.1
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysql-connector-python==9.1.0
numpy==2.1.2
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rq==1.16.2
rq-scheduler==0.13.1
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tenacity==9.0.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
xlrd==2.0.1
zope.interface==7.1.1
